make is good plus qu a verif toutes les fonctions et reverif cond bonus

+++ libft/ft_strncmp.c
@@ -6,14 +6,14 @@
 /*   By: mchrispe <mchrispe@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/16 12:40:33 by mchrispe          #+#    #+#             */
-/*   Updated: 2025/04/18 11:51:26 by mchrispe         ###   ########.fr       */
+/*   Updated: 2025/04/18 13:10:27 by mchrispe         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "libft.h"
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s1[i] != '\0' && s2[i] != '\0' && i < n)
